% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleTSsimulation.R
\name{checkMultipleSimulationVLtimeseries}
\alias{checkMultipleSimulationVLtimeseries}
\title{checkMultipleSimulationVLtimeseries}
\usage{
checkMultipleSimulationVLtimeseries(trueAdjMat, adjMat)
}
\arguments{
\item{trueAdjMat}{a groundtruth matrix.}

\item{adjMat}{an inferred matrix.}
}
\value{
This function returns a list of precision \code{prec}, recall \code{rec}, and F1 score \code{F1} of inferred vs. groundtruth matrices.
}
\description{
checkMultipleSimulationVLtimeseries is a support function that can compare two adjacency matrices: groundtruth and inferred matrices.
It re
}
\examples{
## Generate simulation data
#G<-matrix(FALSE,10,10) # groundtruth
#G[1,c(4,7,8,10)]<-TRUE
#G[2,c(5,7,9,10)]<-TRUE
#G[3,c(6,8,9,10)]<-TRUE
#TS <- MultipleSimulationVLtimeseries()
#out<-multipleVLGrangerFunc(TS)
#checkMultipleSimulationVLtimeseries(trueAdjMat=G,adjMat=out$adjMat)

}
