% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{plotTimeSeries}
\usage{
plotTimeSeries(X, Y, strTitle = "Time Series Plot", TSnames)
}
\arguments{
\item{X}{is a 1st numerical time series}

\item{Y}{is a 2nd numerical time series. If it is not supplied, the function plots only \code{X}.}

\item{strTitle}{is a string of the plot title}

\item{TSnames}{is a list of legend of \code{X,Y} where TSnames[1] is a legend of \code{X} and  TSnames[2] is a legend of \code{Y}.}
}
\value{
This function returns an object of ggplot class.
}
\description{
plotTimeSeries is a function for visualizing time series
}
\examples{
# Generate simulation data
TS <- SimpleSimulationVLtimeseries()
# Run the function
plotTimeSeries(Y=TS$Y,X=TS$X)

}
