% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSNANNearestNeighborPropagation.R
\name{TSNANNearestNeighborPropagation}
\alias{TSNANNearestNeighborPropagation}
\title{TSNANNearestNeighborPropagation}
\usage{
TSNANNearestNeighborPropagation(X)
}
\arguments{
\item{X}{is a T-by-D matrix numerical time series}
}
\value{
This function returns a list of following relation variables below.

\item{Xout}{ is a T-by-D matrix numerical time series that all NAN have been filled with nearest real values.}
}
\description{
TSNANNearestNeighborPropagation is a function that fills NA values with nearest real values in the past ( or future if the first position of time series is NA), for time series \code{X}.
}
\examples{
# Load example data

z<-1:20
z[2:5]<-NA
z<-TSNANNearestNeighborPropagation(z)

}
